roads-interactive: 4 — repair the review's four findings
Adversarial pass on stage 4 returned no blockers and four findings, all acted on:

- `applyEdit`/`clampToBound`/`isAdmissibleDrag`/`chordLength`/`residentTileCount`
  move to a new `src/editPure.ts` with no `@dylanebert/shallot` import anywhere in
  its transitive graph. `edit.ts` re-exports them, so `boot.ts` is unchanged.
  The old comment claiming they were import-free was false: ES imports hoist, so
  importing them from `./edit` pulled the whole device-bound graph — the exact
  Node >=26 hazard Approach stage 4 names. Fixed by making the claim true rather
  than by softening it.
- `__roadsEdit`'s docblock claimed a refusal the bridge does not implement
  (a clamp that leaves the document unchanged). The code is right — the spec asks
  for a length-band refusal only — so the comment now names just that.
- the device handle-y tolerance drops from 0.5 m to `toBeCloseTo(expectedY, 4)`:
  both sides are `heightAtCpu` under `makePermutation(1337)`, so the only
  admissible difference is f32 rounding, and a 0.5 m window would have passed on
  a stale position or a wrong permutation.
- `__roadsVertexFingerprint` becomes an order-sensitive FNV-1a over every word
  instead of a sum, which was invariant under changes that cancel — Validation
  asks for byte-identical.

Gates: suite 363 pass / 0 fail / 17 files, `bun check` exit 0, `bun run format`
clean, device gate exit 0 on `apple metal-3` (re-run after the repair, since R3
and R4 are only exercised there).
